fix(#1401): move A2A FastAPI router from brick to server layer
Per LEGO architecture, API routers belong in the server layer, not in
bricks. Move router.py and create_a2a_router() factory from
nexus.bricks.a2a to nexus.server.api.v2.routers.a2a. Business logic
(handlers, streaming, task_manager) stays in the brick. Update all
consumer imports (fastapi_server.py, 3 e2e test files).
