+++ tests/e2e/self_contained/test_a2a_auth.py
@@ -15,7 +15,7 @@ import pytest
 from fastapi import FastAPI
 from starlette.testclient import TestClient
 
-from nexus.bricks.a2a.router import build_router
+from nexus.server.api.v2.routers.a2a import build_router
 
 # ======================================================================
 # Fixtures
